vm breakdown
