+++ vm/src/lib.rs
@@ -77,9 +77,6 @@ pub mod types;
 pub mod utils;
 pub mod version;
 mod vm;
-mod vm_new;
-mod vm_object;
-mod vm_ops;
 
 mod pyobject {
     pub use super::_pyobject::*;
