+++ vm/src/vm/vm_object.rs
