+++ vm/src/vm/vm_ops.rs
